Issue #171 - implement preflight checks for survey groups, surveys and questions

+++ Dashboard/app/js/lib/views/surveys/survey-details-views.js
@@ -236,27 +236,41 @@ FLOW.QuestionGroupItemView = FLOW.View.extend({
 	}.property('FLOW.selectedControl.selectedForCopyQuestionGroup'),
 
 	// execute group delete
-	// TODO should this be allowed when questions are present?
 	deleteQuestionGroup: function() {
-		var qgDeleteId, questionGroup, questionsGroupsInSurvey, sId, qgOrder;
+		var qgDeleteId, questionGroup, questionsGroupsInSurvey, sId, qgOrder, questionsInGroup;
 		qgDeleteId = this.content.get('keyId');
 		sId = this.content.get('surveyId');
 		questionGroup = FLOW.store.find(FLOW.QuestionGroup, qgDeleteId);
 		qgOrder = questionGroup.get('order');
+		questionsInGroup = FLOW.store.filter(FLOW.Question,function (item){
+			return(item.get('questionGroupId') == qgDeleteId);
+		});
+
+		if (questionsInGroup.get('content').length > 0){
+			FLOW.dialogControl.set('activeAction', "ignore");
+			FLOW.dialogControl.set('header', Ember.String.loc('_cannot_delete_questiongroup'));
+			FLOW.dialogControl.set('message', Ember.String.loc('_cannot_delete_questiongroup_text'));
+			FLOW.dialogControl.set('showCANCEL', false);
+			FLOW.dialogControl.set('showDialog', true);
+			return;
+		}
+
+		// if we are here, we can safely delete
 		questionGroup.deleteRecord();
 		// restore order
 		questionGroupsInSurvey = FLOW.store.filter(FLOW.QuestionGroup, function(item) {
-	        return(item.get('surveyId') == sId);
-	      });
-		
+			return(item.get('surveyId') == sId);
+		});
+
 		questionGroupsInSurvey.forEach(function(item) {
 			if (item.get('order') > qgOrder) {
 				item.set('order',item.get('order')-1);
 			}
 		});
-		
+
 		FLOW.selectedControl.selectedSurvey.set('status', 'NOT_PUBLISHED');
 		FLOW.store.commit();
+	
 	},
 
 	// insert group
